+++ protected/humhub/modules/admin/views/setting/caching.php
@@ -1,8 +1,8 @@
 <?php
 
 use humhub\modules\admin\models\forms\CacheSettingsForm;
-use humhub\modules\ui\form\widgets\ActiveForm;
-use humhub\widgets\Button;
+use humhub\widgets\bootstrap\Button;
+use humhub\widgets\form\ActiveForm;
 
 /* @var $cacheTypes [] */
 /* @var $model CacheSettingsForm */