+++ protected/humhub/modules/comment/widgets/EditLink.php
@@ -8,7 +8,7 @@
 
 namespace humhub\modules\comment\widgets;
 
-use humhub\libs\Html;
+use humhub\helpers\Html;
 use humhub\modules\comment\models\Comment;
 use humhub\modules\ui\icon\widgets\Icon;
 use humhub\modules\ui\menu\WidgetMenuEntry;