+++ protected/humhub/modules/file/widgets/views/showFiles.php
@@ -1,12 +1,12 @@
 <?php
 
 use humhub\components\ActiveRecord;
+use humhub\helpers\Html;
+use humhub\helpers\ThemeHelper;
 use humhub\modules\file\converter\PreviewImage;
 use humhub\modules\file\libs\FileHelper;
 use humhub\modules\file\widgets\FilePreview;
-use humhub\modules\ui\view\helpers\ThemeHelper;
 use humhub\widgets\JPlayerPlaylistWidget;
-use yii\helpers\Html;
 
 /* @var  $previewImage PreviewImage */
 /* @var  $files \humhub\modules\file\models\File[] */
@@ -33,25 +33,28 @@ $nbFiles = count($files);
 $isFluid = ThemeHelper::isFluid();
 
 // Image height is defined in file.less and matches this template
-$mobileBsColumns = 6;
-$desktopBsColumns = $isFluid ? 3 : 4;
+$bsColumns = 6;
+$bsColumnsMd = $isFluid ? 4 : 6;
+$bsColumnsLg = $isFluid ? 3 : 4;
 if ($nbFiles === 1) {
-    $mobileBsColumns = 12;
-    $desktopBsColumns = $isFluid ? 4 : 6;
+    $bsColumns = 12;
+    $bsColumnsMd = $isFluid ? 6 : 12;
+    $bsColumnsLg = $isFluid ? 4 : 6;
 }
 if ($nbFiles === 2) {
-    $desktopBsColumns = $isFluid ? 4 : 6;
+    $bsColumnsMd = 6;
+    $bsColumnsLg = $isFluid ? 4 : 6;
 }
-$fullWidthColumnClass = 'col-media col-xs-12 col-sm-12 col-md-12';
-$galleryColumnClass = 'col-media col-xs-' . $mobileBsColumns . ' col-sm-' . $desktopBsColumns . ' col-md-' . $desktopBsColumns;
+$fullWidthColumnClass = 'col-media col-12';
+$galleryColumnClass = 'col-media col-' . $bsColumns . ' col-lg-' . $bsColumnsMd . ' col-xl-' . $bsColumnsLg;
 ?>
 
 <?php if ($nbFiles > 0): ?>
     <!-- hideOnEdit mandatory since 1.2 -->
-    <div class="hideOnEdit">
+    <div class="container hideOnEdit">
         <!-- Show Images as Thumbnails -->
         <?php if ($showPreview): ?>
-            <div class="post-files clearfix" id="post-files-<?= $object->getUniqueId() ?>">
+            <div class="post-files row" id="post-files-<?= $object->getUniqueId() ?>">
                 <?php if (!empty($audios)): ?>
                     <div class="<?= $fullWidthColumnClass ?>">
                         <?= JPlayerPlaylistWidget::widget(['playlist' => $audios]) ?>