+++ protected/humhub/modules/topic/controllers/ContentTopicController.php
@@ -9,7 +9,7 @@
 
 namespace humhub\modules\topic\controllers;
 
-use humhub\widgets\ModalClose;
+use humhub\widgets\modal\ModalClose;
 use humhub\components\Controller;
 use humhub\modules\content\models\Content;
 use humhub\modules\topic\models\forms\ContentTopicsForm;