+++ protected/humhub/modules/user/notifications/Followed.php
@@ -8,9 +8,9 @@
 
 namespace humhub\modules\user\notifications;
 
-use Yii;
-use yii\bootstrap\Html;
+use humhub\helpers\Html;
 use humhub\modules\notification\components\BaseNotification;
+use Yii;
 use yii\db\IntegrityException;
 
 /**