+++ protected/humhub/modules/web/pwa/widgets/SiteIcon.php
@@ -8,11 +8,11 @@
 
 namespace humhub\modules\web\pwa\widgets;
 
+use humhub\components\View;
+use humhub\components\Widget;
 use humhub\modules\file\libs\FileHelper;
-use humhub\modules\ui\view\components\View;
 use Imagine\Image\Box;
 use Yii;
-use humhub\components\Widget;
 use yii\base\ErrorException;
 use yii\base\Exception;
 use yii\imagine\Image;