+++ CHANGELOG.md
@@ -3,6 +3,7 @@ HumHub Changelog
 
 1.18.0 (Unreleased - Develop)
 ----------------------------
+- Enh #7029: Add Altcha captcha provider for an alternative to the default Yii captcha
 - Enh #7412: Confusing message after creating an account
 - Enh #7425: Installer Wording update due to `.env` configuration